+++ test/stream/main.c
@@ -1433,10 +1433,10 @@ DECLARE_TEST(stream, readwrite_swap) {
 	EXPECT_EQ_MSGFORMAT(read_uint8, expect_uint8, "read uint8 did not swap as expected (%02x)", read_uint8);
 	EXPECT_EQ_MSGFORMAT(read_int16, expect_int16, "read int16 did not swap as expected (%04x)", read_int16);
 	EXPECT_EQ_MSGFORMAT(read_uint16, expect_uint16, "read uint16 did not swap as expected (%04x)", read_uint16);
-	EXPECT_EQ_MSGFORMAT(read_int32, 0x78563412, "read int32 did not swap as expected (%08x)", read_int32);
+	EXPECT_EQ_MSGFORMAT(read_int32, 0x78563412, "read int32 did not swap as expected (%08x)", (uint32_t)read_int32);
 	EXPECT_EQ_MSGFORMAT(read_uint32, 0x32547698UL, "read uint32 did not swap as expected (%04x)", read_uint32);
 	EXPECT_EQ_MSGFORMAT(read_int64, (int64_t)0xefcdab9078563412LL, "read int64 did not swap as expected (%" PRIx64 ")",
-	                    read_int64);
+	                    (uint64_t)read_int64);
 	EXPECT_EQ_MSGFORMAT(read_uint64, 0xefcdab9078563412ULL, "read uint64 did not swap as expected (%" PRIx64 ")",
 	                    read_uint64);
 	EXPECT_EQ_MSGFORMAT(read_uint128.word[0], 0xefcdab9078563412ULL,
@@ -1495,10 +1495,10 @@ DECLARE_TEST(stream, readwrite_swap) {
 	EXPECT_EQ_MSGFORMAT(read_uint8, expect_uint8, "read uint8 did not swap as expected (%02x)", read_uint8);
 	EXPECT_EQ_MSGFORMAT(read_int16, expect_int16, "read int16 did not swap as expected (%04x)", read_int16);
 	EXPECT_EQ_MSGFORMAT(read_uint16, expect_uint16, "read uint16 did not swap as expected (%04x)", read_uint16);
-	EXPECT_EQ_MSGFORMAT(read_int32, 0x78563412, "read int32 did not swap as expected (%08x)", read_int32);
+	EXPECT_EQ_MSGFORMAT(read_int32, 0x78563412, "read int32 did not swap as expected (%08x)", (uint32_t)read_int32);
 	EXPECT_EQ_MSGFORMAT(read_uint32, 0x32547698UL, "read uint32 did not swap as expected (%04x)", read_uint32);
 	EXPECT_EQ_MSGFORMAT(read_int64, (int64_t)0xefcdab9078563412LL, "read int64 did not swap as expected (%" PRIx64 ")",
-	                    read_int64);
+	                    (uint64_t)read_int64);
 	EXPECT_EQ_MSGFORMAT(read_uint64, 0xefcdab9078563412ULL, "read uint64 did not swap as expected (%" PRIx64 ")",
 	                    read_uint64);
 	EXPECT_EQ_MSGFORMAT(read_uint128.word[0], 0xefcdab9078563412ULL,
